Add files via upload
This is a single python script, which combines all the scripts into one.

Creating individual python scripts for login, getting response from server, getting mission response, etc. resulted in a forbidden 403 error. This is because in order to get the response from the server,  it should identify the client, which is done only when the login is done. 
This forced me to include initial login python scripts in all the individual python files created. In order to solve this issue,all the individual python scripts are combined into a single python script of commands.py .This works well.